requirements.txt
python-dotenv
pandas
numpy
requests
flask
sqlalchemy
pytest
setuptools
pymongo
certifi 
scikit-learn
# pymongo[srv]
# -e .